# openapi.json
{"openapi":"3.1.0","info":{"title":"AI Orchestrator","version":"0.1.0"},"paths":{"/books/run":{"post":{"tags":["books"],"summary":"Books Run","operationId":"books_run_books_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunBookRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/{book}/current":{"get":{"tags":["books"],"summary":"Books Current","operationId":"books_current_books__book__current_get","parameters":[{"name":"book","in":"path","required":true,"schema":{"type":"string","title":"Book"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/{book}/status":{"get":{"tags":["books"],"summary":"Books Status","operationId":"books_status_books__book__status_get","parameters":[{"name":"book","in":"path","required":true,"schema":{"type":"string","title":"Book"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/agent/run":{"post":{"tags":["books"],"summary":"Agent Run","operationId":"agent_run_books_agent_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentRunRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunJobResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/jobs/{job_id}":{"get":{"tags":["books"],"summary":"Get Job","operationId":"get_job_books_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/agent/{book}/draft":{"get":{"tags":["books"],"summary":"Agent Draft","operationId":"agent_draft_books_agent__book__draft_get","parameters":[{"name":"book","in":"path","required":true,"schema":{"type":"string","title":"Book"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/agent/{book}/state":{"get":{"tags":["books"],"summary":"Agent State","operationId":"agent_state_books_agent__book__state_get","parameters":[{"name":"book","in":"path","required":true,"schema":{"type":"string","title":"Book"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/agent/{book}/log":{"get":{"tags":["books"],"summary":"Agent Log","operationId":"agent_log_books_agent__book__log_get","parameters":[{"name":"book","in":"path","required":true,"schema":{"type":"string","title":"Book"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AgentRunRequest":{"properties":{"book":{"type":"string","minLength":1,"title":"Book"},"topic":{"type":"string","minLength":1,"title":"Topic"},"target_words":{"type":"integer","exclusiveMinimum":0.0,"title":"Target Words","default":1200},"chunk_words":{"type":"integer","exclusiveMinimum":0.0,"title":"Chunk Words","default":300},"model":{"type":"string","title":"Model","default":"gpt-4.1-mini"},"max_output_tokens":{"type":"integer","exclusiveMinimum":0.0,"title":"Max Output Tokens","default":1200},"tail_chars":{"type":"integer","exclusiveMinimum":0.0,"title":"Tail Chars","default":6000},"job_timeout_sec":{"type":"integer","exclusiveMinimum":0.0,"title":"Job Timeout Sec","default":3600}},"type":"object","required":["book","topic"],"title":"AgentRunRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobInfo":{"properties":{"job_id":{"type":"string","title":"Job Id"},"status":{"type":"string","enum":["QUEUED","RUNNING","DONE","ERROR"],"title":"Status"},"kind":{"type":"string","title":"Kind"},"book":{"type":"string","title":"Book"},"created_at":{"type":"number","title":"Created At"},"started_at":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Started At"},"finished_at":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Finished At"},"exit_code":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Exit Code"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"paths":{"additionalProperties":{"type":"string"},"type":"object","title":"Paths"},"stdout_tail":{"type":"string","title":"Stdout Tail","default":""},"stderr_tail":{"type":"string","title":"Stderr Tail","default":""},"meta":{"additionalProperties":true,"type":"object","title":"Meta"}},"type":"object","required":["job_id","status","kind","book","created_at"],"title":"JobInfo"},"RunBookRequest":{"properties":{"book":{"type":"string","minLength":1,"title":"Book"},"delta":{"type":"integer","exclusiveMinimum":0.0,"title":"Delta","default":120},"model":{"type":"string","title":"Model","default":"gpt-4.1-mini"},"max_output_tokens":{"type":"integer","exclusiveMinimum":0.0,"title":"Max Output Tokens","default":1200},"prompt_text":{"type":"string","minLength":1,"title":"Prompt Text"}},"type":"object","required":["book","prompt_text"],"title":"RunBookRequest"},"RunJobResponse":{"properties":{"job_id":{"type":"string","title":"Job Id"},"status":{"type":"string","enum":["QUEUED","RUNNING","DONE","ERROR"],"title":"Status"},"paths":{"additionalProperties":{"type":"string"},"type":"object","title":"Paths"}},"type":"object","required":["job_id","status","paths"],"title":"RunJobResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}